Update Project : Improved package structure
