+++ app/src/main/java/com/yong/blog/detail/ui/DetailScreen.kt
@@ -1,12 +1,12 @@
-package com.yong.blog.screen.list
+package com.yong.blog.detail.ui
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 
 @Composable
-fun ListScreen(
-    onNavigateToDetail: () -> Unit,
+fun DetailScreen(
+    onNavigateToList: () -> Unit,
     onNavigateToMain: () -> Unit
 ) {
-    Text("List")
+    Text("Detail")
 }
