+++ app/src/main/java/com/yong/blog/navigation/RouteDefinition.kt
@@ -1,4 +1,4 @@
-package com.yong.blog.screen
+package com.yong.blog.navigation
 
 sealed class RouteDefinition(val route: String) {
     object Main: RouteDefinition("main")
